requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gspread==5.12.4
oauth2client==4.1.3
gunicorn==21.2.0
python-dotenv==1.0.0 